+++ neocarta/ingest/lpg/__init__.py
@@ -1,0 +1,5 @@
+"""Neo4j LPG Ingest Module."""
+
+from .load import Neo4jLPGLoader
+
+__all__ = ["Neo4jLPGLoader"]